+++ src/Network/Network/Models/Cortex/PSExpressRouteConnection.cs
@@ -31,6 +31,9 @@ namespace Microsoft.Azure.Commands.Network.Models
         [Ps1Xml(Label = "Internet Security Enabled", Target = ViewControl.Table)]
         public bool EnableInternetSecurity { get; set; }
 
+        [Ps1Xml(Label = "ExpressRoute GatewayBypass Enabled", Target = ViewControl.Table)]
+        public bool ExpressRouteGatewayBypass { get; set; }
+
         public PSRoutingConfiguration RoutingConfiguration { get; set; }
 
         [Ps1Xml(Label = "Provisioning State", Target = ViewControl.Table)]