+++ src/Network/Network/ChangeLog.md
@@ -27,6 +27,9 @@
     - Added parameter -SkuTier
     - Added parameter -SkuName and made Sku as Alias for this
     - Removed parameter -Sku
+* Updated ExpressRouteConnection:
+    - Added parameter `-ExpressRouteGatewayBypass` in `New-AzExpressRouteConnection`
+    - Added parameter `-ExpressRouteGatewayBypass` in `Set-AzExpressRouteConnection`
     
 ## Version 3.5.0
 * Added Office365 Policy to VPNSite Resource